move sampleReport to __fixture__
